load_student() function allows the user to extract the PISA student scores for any desired year from 2000-2018

load_student() function was created to extract the data of student's scores in any years from 2000-2018, the function requires any of the year as it argument or a string "all" that will return all the PISA scores of the students from the years 2000-2018.

load_student(year = "2000")

Arguments

year

is the required parameter for the function to display the dataset the user wants to view the PISA scores for the selected year else the entire student data will be available to the user

Value

A dataset of PISA scores of students that took the test in the selected year as per user from the years 2000-2018

Examples

if (FALSE) { # \dontrun{
library(learningtower)
student_all <- load_student("all")
student_2000 <- load_student("2000")
} # }