Subset of the School data available for the years 2000-2022 from the PISA OECD database

A subset data containing school weight and other information from the triennial testing of 15 year olds around the globe. Original data available from https://www.oecd.org/en/about/programmes/pisa/pisa-data.html.

Format

A tibble of the following variables

year: Year of the PISA data. Integer.

country: Country 3 character code. Note that some regions/territories are coded as country for ease of input. Chracter.

school_id: The school identification number, unique for each country and year combination. Character.

fund_gov: Percentage of total funding for school year from government. Numeric.

fund_fees: Percentage of total funding for school year from student fees or school charges paid by parents. Numeric.

fund_donation: Percentage of total funding for school year from benefactors, donations, bequests, sponsorship, parent fundraising. Numeric.

enrol_boys: Number of boys in the school. Numeric.

enrol_girls: Number of girls in the school. Numeric.

stratio: Student-Teacher ratio. Numeric.

public_private: Is the school a public or private school. Factor.

staff_shortage: Shortage of staff. Numeric.

sch_wgt: The final survey weight score for the schools. Numeric.

school_size: The school size. Numeric.