Student data from 2000-2018 PISA OECD data

A dataset containing scores and other information from the triennial testing of 15 year olds around the globe. Original data available from https://www.oecd.org/pisa/data/.

Format

A tibble of the following variables

year: Year of the PISA data. Factor.

country: Country 3 character code. Factor.

school_id: The school identification number, unique for each country and year combination. Factor.

student_id: The student identification number, unique for each school, country and year combination. Factor.

mother_educ: Highest level of mother's education. Ranges from "less than ISCED1" to "ISCED 3A". Factor. Note that in 2000, all entries are missing.

father_educ: Highest level of father's education. Ranges from "less than ISCED1" to "ISCED 3A". Factor. Note that in 2000, all entries are missing.

gender: Gender of the student. Only "male" and "female" are recorded. Factor.

computer: Possession of computer. Only "yes" and "no" are recorded. Factor.

internet: Access to internet. Only "yes" and "no" are recorded. Factor.

math: Simulated score in mathematics. Numeric.

read: Simulated score in reading. Numeric.

science: Simulated score in science. Numeric.

stu_wgt: The final survey weight score for the student score. Numeric.

desk: Possession of desk to study at. Only "yes" and "no" are recorded. Factor.

room: Possession of a room of your own. Only "yes" and "no" are recorded. Factor.

dishwasher: Possession of a dishwasher. Only "yes" and "no" are recorded. Factor. Note that in 2015 and 2018, all entries are missing.

television: Number of televisions. "0", "1", "2" are code for no, one and two TVs in the house. "3+" codes for three or more TVs. Factor. Note that in 2003, all entries are missing.

computer_n: Number of computers. "0", "1", "2" are code for no, one and two computers in the house. "3+" codes for three or more computers. Factor. Note that in 2003, all entries are missing.

car: Number of cars. "0", "1", "2" are code for no, one and two cars in the house. "3+" codes for three or more cars Factor. Note that in 2003, all entries are missing.

book: Number of books. Factor. Note that encodings are different in 2000 and 2003 compare to all other years. Factor. Evalute table(student$book, student$year) for a demo.

wealth: Family wealth. Numeric. Note that in 2003, all entries are missing.

escs: Index of economic, social and cultural status. Numeric.