student_data function allows the user to extract the PISA scores for any desired year from 2000-2018

student_data function was created to extract the data of student's scores in any years from 2000-2018, the function requires any of the year as it argument or a string "all" that will return the all the PISA scores of the student from 2000-2018.

student_data(year)

Arguments

year
is the required parameter for the function to display the dataset the user wants to view the PISA scores for the selected year else the entire student data will be available to the user

Value

A dataset of PISA scores of students that took the test in the selected year as per user from 2000-2018

Examples

student_data(2018)

#> # A tibble: 612,004 × 22
#>    year  country school_id student_id mother_educ father_educ gender computer
#>    <fct> <fct>   <fct>     <fct>      <fct>       <fct>       <fct>  <fct>   
#>  1 2018  ALB     800002    800251     ISCED 3A    ISCED 3A    male   yes     
#>  2 2018  ALB     800002    800402     ISCED 2     ISCED 2     male   yes     
#>  3 2018  ALB     800002    801902     ISCED 2     ISCED 2     female no      
#>  4 2018  ALB     800002    803546     ISCED 2     ISCED 2     male   no      
#>  5 2018  ALB     800002    804776     ISCED 2     ISCED 3A    male   yes     
#>  6 2018  ALB     800002    804825     ISCED 2     ISCED 2     female yes     
#>  7 2018  ALB     800002    804983     NA          NA          female NA      
#>  8 2018  ALB     800002    805287     ISCED 2     ISCED 2     male   no      
#>  9 2018  ALB     800002    805601     ISCED 2     ISCED 2     female yes     
#> 10 2018  ALB     800002    806295     NA          NA          female NA      
#> # … with 611,994 more rows, and 14 more variables: internet <fct>, math <dbl>,
#> #   read <dbl>, science <dbl>, stu_wgt <dbl>, desk <fct>, room <fct>,
#> #   dishwasher <fct>, television <fct>, computer_n <fct>, car <fct>,
#> #   book <fct>, wealth <dbl>, escs <dbl>

student_data("all")

#> # A tibble: 2,929,621 × 22
#>    year  country school_id student_id mother_educ father_educ gender computer
#>    <fct> <fct>   <fct>     <fct>      <fct>       <fct>       <fct>  <fct>   
#>  1 2000  ALB     1001      1          NA          NA          female NA      
#>  2 2000  ALB     1001      3          NA          NA          female NA      
#>  3 2000  ALB     1001      6          NA          NA          male   NA      
#>  4 2000  ALB     1001      8          NA          NA          female NA      
#>  5 2000  ALB     1001      11         NA          NA          female NA      
#>  6 2000  ALB     1001      12         NA          NA          female NA      
#>  7 2000  ALB     1001      17         NA          NA          male   NA      
#>  8 2000  ALB     1001      20         NA          NA          female NA      
#>  9 2000  ALB     1001      21         NA          NA          female NA      
#> 10 2000  ALB     1001      22         NA          NA          female NA      
#> # … with 2,929,611 more rows, and 14 more variables: internet <fct>,
#> #   math <dbl>, read <dbl>, science <dbl>, stu_wgt <dbl>, desk <fct>,
#> #   room <fct>, dishwasher <fct>, television <fct>, computer_n <fct>,
#> #   car <fct>, book <fct>, wealth <dbl>, escs <dbl>

student_data(2009)

#> # A tibble: 515,958 × 22
#>    year  country school_id student_id mother_educ father_educ gender computer
#>    <fct> <fct>   <fct>     <fct>      <fct>       <fct>       <fct>  <fct>   
#>  1 2009  ALB     00019     00518      ISCED 1     ISCED 3A    female no      
#>  2 2009  ALB     00019     00520      ISCED 3A    ISCED 3B, C female yes     
#>  3 2009  ALB     00019     00507      NA          ISCED 3A    female no      
#>  4 2009  ALB     00019     00510      ISCED 3A    NA          female yes     
#>  5 2009  ALB     00019     00506      ISCED 3A    ISCED 3A    female yes     
#>  6 2009  ALB     00019     00496      ISCED 2     ISCED 3A    female no      
#>  7 2009  ALB     00019     00521      ISCED 3A    NA          female no      
#>  8 2009  ALB     00019     00517      ISCED 2     ISCED 2     female no      
#>  9 2009  ALB     00019     00516      ISCED 2     ISCED 3A    female no      
#> 10 2009  ALB     00019     00522      ISCED 3B, C ISCED 3A    female no      
#> # … with 515,948 more rows, and 14 more variables: internet <fct>, math <dbl>,
#> #   read <dbl>, science <dbl>, stu_wgt <dbl>, desk <fct>, room <fct>,
#> #   dishwasher <fct>, television <fct>, computer_n <fct>, car <fct>,
#> #   book <fct>, wealth <dbl>, escs <dbl>