student_data function allows the user to extract the PISA scores for any desired year from 2000-2018

student_data function was created to extract the data of student's scores in any years from 2000-2018, the function requires any of the year as it argument or a string "all" that will return the all the PISA scores of the student from 2000-2018.

student_data(year)

Arguments

year
is the required parameter for the function to display the dataset the user wants to view the PISA scores for the selected year else the entire student data will be available to the user

Value

A dataset of PISA scores of students that took the test in the selected year as per user from 2000-2018

Examples

student_data(2018)

#> Error in student_data(2018): object 'x' not found

student_data("all")

#> Error in student_data("all"): object 'x' not found

student_data(2009)

#> Error in student_data(2009): object 'x' not found